+++ CONTROLLER/SituacaoCadastral.cs
@@ -6,11 +6,11 @@ namespace Teu_Assistente_HABITACAO
     internal class SituacaoCadastral : BDAgendamento
     {
         public string situacao { get; set; }
-        public void definirSituacao(int cpf, string situacao)
+        public void definirSituacao(long cpf, string situacao)
         {
             this.setAgendamento(cpf);
         }
-        public void consultarSituacao(int cpf)
+        public void consultarSituacao(long cpf)
         {
             this.getAgendamento(cpf);
         }
